infogob/requirements.txt
beautifulsoup4==4.8.0
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.1.10
django-allauth==0.39.0
django-rest-auth==0.9.3
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
idna==2.8
oauthlib==3.1.0
psycopg2-binary==2.7.7
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2018.9
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
soupsieve==1.9.3
sqlparse==0.3.0
urllib3==1.25.3